LayoutTests/storage/indexeddb/transaction-read-only-expected.txt

Test read-only transactions in IndexedDB.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open('transaction-read-only')
openSuccess():
db = event.target.result
result = db.setVersion('version 1')
Deleted all object stores.
store = db.createObjectStore('store')
store.put('x', 'y')
trans = db.transaction('store')
Expecting exception from trans.objectStore('store').put('a', 'b')
PASS Exception was thrown.
PASS code is IDBDatabaseException.READ_ONLY_ERR
PASS ename is 'ReadOnlyError'
trans = db.transaction('store')
Expecting exception from trans.objectStore('store').delete('x')
PASS Exception was thrown.
PASS code is IDBDatabaseException.READ_ONLY_ERR
PASS ename is 'ReadOnlyError'
trans = db.transaction('store')
cur = trans.objectStore('store').openCursor()
PASS !event.target.result is false
Expecting exception from event.target.result.delete()
PASS Exception was thrown.
PASS code is IDBDatabaseException.READ_ONLY_ERR
PASS ename is 'ReadOnlyError'
PASS successfullyParsed is true

TEST COMPLETE
